requirements.txt
airtable-python-wrapper==0.15.1
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-appconf==1.0.4
django-imagefit==0.6.2
idna==2.10
Pillow==8.1.0
psycopg2==2.8.6
python-dotenv==0.15.0
python-memcached==1.59
pytz==2021.1
requests==2.25.1
six==1.15.0
sorl-thumbnail==12.7.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
urllib3==1.26.3
